refactor: moved buttons to CurrencySelector.js component and clean
