+++ src/App.js
@@ -1,11 +1,10 @@
 import { useState, useEffect } from 'react'
 import Container from 'react-bootstrap/Container'
 import Jumbotron from 'react-bootstrap/Jumbotron'
-import Button from 'react-bootstrap/Button'
-import ButtonGroup from 'react-bootstrap/ButtonGroup'
 import axios from 'axios'
 import dayjs from 'dayjs'
 import Price from './components/Price'
+import CurrencySelector from './components/CurrencySelector'
 
 dayjs().format()
 
@@ -15,11 +14,8 @@ function App() {
   const [timeData, setTimeData] = useState()
   const [currency, setCurrency] = useState({ code: 'AUD', symbol: '$' })
 
-  const symbolReference = {
-    AUD: '$',
-    USD: '$',
-    EUR: '€',
-    GBP: '£',
+  const handleCurrency = (event) => {
+    setCurrency(event)
   }
 
   useEffect(() => {
@@ -63,75 +59,17 @@ function App() {
     }
   }, [currency])
 
-  // Handler for buttons to set currency and currencySymbol state
-  const clickHandler = (props) => {
-    let selectedCurrency = props.target.innerText
-    setCurrency({
-      code: selectedCurrency,
-      symbol: symbolReference[selectedCurrency],
-    })
-  }
-
   return (
     <Container className="mt-5">
       <Jumbotron className="text-center">
         <h1>BTC Tracker</h1>
-        {/* <Container
-          style={{ height: '180px' }}
-          className="d-flex align-items-center justify-content-center"
-        >
-          {isLoading ? (
-            <Spinner animation="border" className="text-primary" />
-          ) : (
-            <div>
-              <p>
-                <h2 className="display-1">
-                  {currency.symbol}
-                  {priceData}
-                </h2>
-              </p>
-              <p className="text-secondary">as at {timeData}</p>
-            </div>
-          )}
-        </Container> */}
-
         <Price
           displayPrice={priceData}
           displayTime={timeData}
           displaySymbol={currency.symbol}
           isLoading={isLoading}
         />
-
-        <ButtonGroup>
-          <Button
-            onClick={clickHandler}
-            disabled={currency.code === 'AUD'}
-            variant="primary"
-          >
-            AUD
-          </Button>
-          <Button
-            onClick={clickHandler}
-            disabled={currency.code === 'USD'}
-            variant="primary"
-          >
-            USD
-          </Button>
-          <Button
-            onClick={clickHandler}
-            disabled={currency.code === 'EUR'}
-            variant="primary"
-          >
-            EUR
-          </Button>
-          <Button
-            onClick={clickHandler}
-            disabled={currency.code === 'GBP'}
-            variant="primary"
-          >
-            GBP
-          </Button>
-        </ButtonGroup>
+        <CurrencySelector currency={currency} handleCurrency={handleCurrency} />
       </Jumbotron>
     </Container>
   )
